Added the option to specify any class as unlabeled for augmenting
augment_and_save() now has an extra input argument, nanID. This is
exactly like the nanID argument in load_training_data(), since it's
used to specify which class is for all unlabeled pixels that are
to be ignored. The augmenter previously assumed that the last class
in its pixelLabelDatastore was for unlabeled pixels, however this
new approach makes the function more flexible & easier to use with
other indexed-color datasets.