Created new approach to augmentation & training
aug_learn_script uses an ImageDataAugmenter on each of the raw
input patches instead of pre-augmented images like the other
script does. This doesn't seem to increase the training efficiency
or accuracy much, but it does save space & time when creating the
training patches.

Also added some new tracings, most notably some sections of RLG_80.
This model does a much better job when segmenting all of #80 as a
result.