Created SVM + Script for identifying shell segments
Scans through each output image from the segmenting UNet and
deletes all crystal segments that are not identified to be shells.
The SVM was created through MATLAB's Classification app, and was
able to remove a significant amount of non-shell components from
a small test set while minimizing the number of true shells
that get mistakenly removed.

However, many non-shell components get marked as shells and are
kept, so the output of this script should be manually looked over
for any mistakes.